Moved stats code back to the generate-report command.
